# api/openapi.yaml
openapi: 3.1.0
info:
  title: RemitX Anchor API
  version: "0.1.0"
  description: |
    RemitX Stellar anchor API implementing SEP-1, SEP-10, SEP-12, and SEP-31
    for trustless cross-border remittances on Stellar.

    ## Authentication

    Most endpoints require a **SEP-10 JWT** obtained from `POST /auth`:

    ```
    Authorization: Bearer <jwt>
    ```

    The `/api/v1/escrow/:id/release` and `/api/v1/escrow/:id/dispute` endpoints
    additionally require the JWT to be signed with an **Ed25519 key** (full
    SEP-10 Ed25519 verification, not the shared-secret variant).

    Content / tier management endpoints authenticate via the
    `X-Stellar-Account` header (Stellar G… public key).

    The admin reconciliation endpoint requires a separate admin JWT with
    `role: "admin"`.

    ## Idempotency

    All mutating endpoints (POST / PUT / PATCH / DELETE) **must** include an
    `Idempotency-Key` header (UUID v4) so that retries are safe. The server
    replays the original response for duplicate keys.

  contact:
    name: RemitX Engineering
    email: dev@remitx.io
    url: https://discord.gg/remitx
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://api.remitx.io
    description: Production
  - url: http://localhost:8000
    description: Local development

# ---------------------------------------------------------------------------
# Security schemes
...
components:
  securitySchemes:
    sep10JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT obtained from `POST /auth` after completing the SEP-10
        challenge/response flow. Valid for 24 hours.

    adminJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Admin JWT with `role: "admin"` claim. Issued out-of-band to
        operations staff.

    stellarAccount:
      type: apiKey
      in: header
      name: X-Stellar-Account
      description: >
        Stellar Ed25519 public key (G… address) of the authenticated creator
        or fan. Used by content / tier management endpoints.

  # -------------------------------------------------------------------------
  # Reusable headers
  # -------------------------------------------------------------------------
...
  parameters:
...
    TransactionId:
      name: id
      in: path
      required: true
      description: UUID of the SEP-31 transaction.
      schema:
        type: string
        format: uuid
        example: "7c9e6679-7425-40de-944b-e07fc1f90ae7"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message.
          example: "escrow not found"
        detail:
          type: string
          description: Optional extended detail (admin / server errors only).
          example: "connection refused: postgres:5432"

    # ── SEP-10 ──────────────────────────────────────────────────────────────
...
    Sep31TransactionRequest:
      type: object
      required: [asset_code, amount, sender_id, receiver_id]
      properties:
        asset_code:
          type: string
          example: "USDC"
        asset_issuer:
          type: string
          description: Stellar issuer account. Optional for USDC on testnet.
          example: "GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN"
        amount:
          type: string
          description: Decimal amount to transfer.
          example: "100.00"
        sender_id:
          type: string
          format: uuid
          description: SEP-12 customer ID of the sending party.
        receiver_id:
          type: string
          format: uuid
          description: SEP-12 customer ID of the receiving party.
        fields:
          type: object
          properties:
            transaction:
              type: object
              properties:
                receiver_account_number:
                  type: string
                  example: "0123456789"
                receiver_routing_number:
                  type: string
                  example: "044151003"
    Sep31TransactionCreatedResponse:
      type: object
      required: [id, stellar_account_id, stellar_memo_type, stellar_memo]
      properties:
        id:
          type: string
          format: uuid
        stellar_account_id:
          type: string
          example: "GBBD47IF6LWK7P7MDEVSCWR7DPUWV3NY3DTQEVFL4NAT4AQH3ZLLFLA5"
        stellar_memo_type:
          type: string
          example: "hash"
        stellar_memo:
          type: string
          example: "yVtBlS1iqk8Qp/A1qKMEb0i+Xs+xr4tqIv2BaoPT8QM="
    Sep31Transaction:
      type: object
      required:
        - id
        - status
        - amount_in
        - amount_in_asset
        - amount_out
        - amount_out_asset
        - amount_fee
        - amount_fee_asset
        - stellar_account_id
        - stellar_memo_type
        - stellar_memo
        - started_at
        - updated_at
        - refunded
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - pending_sender
            - pending_stellar
            - pending_transaction_info_update
            - pending_receiver
            - pending_external
            - completed
            - error
          example: "pending_sender"
        status_eta:
          type: integer
          nullable: true
          description: Estimated seconds until next status change.
        status_message:
          type: string
          nullable: true
        amount_in:
          type: string
          example: "100.00"
        amount_in_asset:
          type: string
          example: "stellar:USDC"
        amount_out:
          type: string
          example: "99.50"
        amount_out_asset:
          type: string
          example: "stellar:USDC"
        amount_fee:
          type: string
          example: "0.50"
        amount_fee_asset:
          type: string
          example: "stellar:USDC"
        stellar_account_id:
          type: string
        stellar_memo_type:
          type: string
        stellar_memo:
          type: string
        started_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
        stellar_transaction_id:
          type: string
          nullable: true
        external_transaction_id:
          type: string
          nullable: true
        refunded:
          type: boolean
        required_info_message:
          type: string
          nullable: true
        required_info_updates:
          type: object
          nullable: true
    Sep31PatchRequest:
      type: object
      required: [fields]
      properties:
        fields:
          type: object
          required: [transaction]
          properties:
            transaction:
              type: object
              additionalProperties:
                type: string
              example:
                receiver_account_number: "0987654321"
    Sep31CallbackRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
          example: "https://sending-anchor.example.com/callbacks/sep31"

    # ── Escrow ──────────────────────────────────────────────────────────────
...
  responses:
    400BadRequest:
      description: Bad request — missing or invalid input.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "sender_account is required"
    401Unauthorized:
      description: Missing or invalid authentication token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Unauthorized"
...
    404NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "escrow not found"
...
paths:
...
  /sep31/transactions:
    post:
      operationId: sep31CreateTransaction
      summary: SEP-31 — initiate a cross-border payment
      description: >
        Creates a new SEP-31 transaction. The response includes the Stellar
        account and memo to which the sending anchor should send funds.
      tags: [SEP-31 Payments]
      security:
        - sep10JWT: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Sep31TransactionRequest"
            example:
              asset_code: "USDC"
              amount: "100.00"
              sender_id: "7c9e6679-7425-40de-944b-e07fc1f90ae7"
              receiver_id: "a3bb189e-8bf9-3888-9912-ace4e6543002"
              fields:
                transaction:
                  receiver_account_number: "0123456789"
      responses:
        "201":
          description: Transaction created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sep31TransactionCreatedResponse"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
  /sep31/transactions/{id}:
    get:
      operationId: sep31GetTransaction
      summary: SEP-31 — get transaction status
      tags: [SEP-31 Payments]
      security:
        - sep10JWT: []
      parameters:
        - $ref: "#/components/parameters/TransactionId"
      responses:
        "200":
          description: Transaction detail.
          content:
            application/json:
              schema:
                type: object
                required: [transaction]
                properties:
                  transaction:
                    $ref: "#/components/schemas/Sep31Transaction"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "404":
          $ref: "#/components/responses/404NotFound"

    patch:
      operationId: sep31PatchTransaction
      summary: SEP-31 — supply missing transaction info
      description: >
        Provides the fields previously flagged as required by the anchor. The
        transaction must be in `pending_transaction_info_update` status.
      tags: [SEP-31 Payments]
      security:
        - sep10JWT: []
      parameters:
        - $ref: "#/components/parameters/TransactionId"
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Sep31PatchRequest"
      responses:
        "200":
          description: Transaction updated.
          content:
            application/json:
              schema:
                type: object
                required: [transaction]
                properties:
                  transaction:
                    $ref: "#/components/schemas/Sep31Transaction"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "404":
          $ref: "#/components/responses/404NotFound"
  /sep31/transactions/{id}/callback:
    put:
      operationId: sep31SetCallback
      summary: SEP-31 — register a status callback URL
      description: >
        Registers a URL that will receive POST notifications when the
        transaction status changes.
      tags: [SEP-31 Payments]
      security:
        - sep10JWT: []
      parameters:
        - $ref: "#/components/parameters/TransactionId"
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Sep31CallbackRequest"
      responses:
        "204":
          description: Callback URL registered.
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "404":
          $ref: "#/components/responses/404NotFound"

  # ---------------------------------------------------------------------------
  # Escrow — payment initiation and lifecycle
  # ---------------------------------------------------------------------------
...